reorg: Add a real query clause processor (mostly), move test binary to a cmd/ subdirectory
