+++ .gitignore
@@ -1,1 +1,1 @@
-querydsl
+cmd/querydsltest/querydsltest
